[coq manager] Remove all unneeded methods
This is a try, to see what the impact is, but looks very good.
