+++ frontend/classic/js/cm-provider.ts
@@ -10,7 +10,7 @@ import localforage from "localforage";
 import $ from 'jquery';
 
 // CM imports
-import CodeMirror from "codemirror";
+import CodeMirror, { Editor } from "codemirror";
 
 import 'codemirror/addon/hint/show-hint.js';
 import 'codemirror/addon/edit/matchbrackets.js';
@@ -32,6 +32,8 @@ import 'codemirror/addon/dialog/dialog.css';
 import '../external/CodeMirror-TeX-input/addon/hint/tex-input-hint.js';
 import './mode/coq-mode.js';
 import { CompanyCoq }  from './addon/company-coq.js';
+
+import { Deprettify } from './deprettify';
 
 /**
  * A Coq sentence, typically ended in dot "."
@@ -108,8 +110,8 @@ export class CmCoqProvider {
     dirty : boolean;
     autosave : any; // result of setTimeout
     autosave_interval : number;
-    editor : CodeMirror.Editor;
-    onChange : (cm, change ) => void;
+    editor : Editor;
+    onChange : (cm : Editor, change) => void;
     onInvalidate : (evt : any ) => void;
     onMouseEnter : (stm, evt : any ) => void;
     onMouseLeave : (stm, evt : any ) => void;
@@ -924,65 +926,6 @@ function pageUpDownOverride(element) {
             }
         }, {capture: true});
 }
-
-/**
- * For HTML-formatted Coq snippets created by coqdoc.
- * This reverses the modifications made during pretty-printing
- * to allow the text to be placed in an editor.
- */
-export class Deprettify {
-    static REPLACES : [RegExp, string][];
-    /**
-     * Remove redundant leading and trailing newlines generated by coqdoc.
-     * @param {HTMLElement} element 
-     */
-    static trim(element) {
-        var c;
-        if ((c = element.firstChild) && Deprettify.isWS(c))
-            element.removeChild(c);
-        if ((c = element.firstChild) && Deprettify.isBR(c))
-            element.removeChild(c);
-        //if ((c = element.firstChild) && Deprettify.isWS(c))
-        //    element.removeChild(c);
-        while ((c = element.lastChild) && 
-                (Deprettify.isWS(c) || Deprettify.isBR(c)))
-            element.removeChild(element.lastChild);
-        return element;
-    }
-
-    /**
-     * @param {ChildNode} node
-     */
-    static isWS(node) {
-        return node.nodeType === Node.TEXT_NODE &&
-               node.nodeValue.match(/^\s*\n$/);
-    }
-
-    /**
-     * @param {ChildNode} node
-     */
-    static isBR(node) {
-        return node.nodeType === Node.ELEMENT_NODE &&
-               node.nodeName === 'BR';
-    }
-
-    /**
-     * Translate back some unicode symbols to their original ASCII.
-     * @param {string} text 
-     */
-    static cleanup(text) {
-        for (let [re, s] of this.REPLACES) text = text.replace(re, s);
-        return text.replace(/^\n/, '');
-    }
-}
-
-/* Safari does not support static members? */
-Deprettify.REPLACES = /** @type {[RegExp, string][]} */ ([
-    [/\xa0/g, ' '], [/⇒/g, '=>'],   [/×/g, '*'],
-    [/→/g, '->'],   [/←/g, '<-'],   [/¬/g, '~'],
-    [/⊢/g, '|-'],   [/\n☐/g, ''],
-    [/∃/g, 'exists']  /* because it is also a tactic... */
-]);
 
 // Local Variables:
 // js-indent-level: 4
